FIX: Error in client site settings JSON (#33640)
Followup 19af83d39e6c06cdc31df2c203623b47bef9c252

In some scenarios, the client_settings_json_uncached
was silently failing and returning "", which caused
problems down the line in mini_racer when this code
was produced:

```
__optInput = {};
__optInput.siteSettings = ;
```

The problem was only happening in rare cases with interaction
between core and other plugins, because the length of the
array we were trying to make a Hash with wasn't an even length.

To address this, we can use filter_map to ensure correct behavior,
and add some specs and also raise errors when parsing client settings
JSON
in specs, so we don't miss this in the future.