FIX: Error in shadowed global settings for deprecated settings (#33642)
Followup 19af83d39e6c06cdc31df2c203623b47bef9c252

There was another edge case here, where this scenario happened:

* A site setting is deprecated _without_ an override
* This site setting is being shadowed by a global setting
* This would cause an argument error when getting the deprecated
  setting value in SiteSetting.client_settings_json_uncached

An example of this setting was `external_system_avatars_enabled`.
The shadowed setting was blowing up because it was missing the
`scoped_to` argument that was added for themeable site settings
in the referenced commit.

This commit fixes the issue and adds a spec to cover this scenario,
followup commit will add stronger handling of shadowed global
settings for theme site settings...this is not allowed.