DEV: Ensure `streamElement` consistency across components (#33638)
This commit introduces the `@streamElement` argument to various
components and modifies its default behavior to `false` in
`cooked-html`. It standardizes the flow, ensuring better control and
accurate context sharing of the `streamElement` property. Adjustments
were also made to decorators and nested post rendering.

Additionally, tests in the `poll` plugin were modified to run with
`glimmer_post_stream_mode` enabled and disabled to ensure the decorators
are working as expected.