FIX: Serialize time objects properly in calendar plugin
Currently, `#event_starts_at` and `#event_ends_at` are not properly
serialized: we just output the string we have in the custom fields,
which is not necessarily a time string JS can deserialize. Until now it
worked by chance.

With Rails 8, the string format changes a bit to include the timezone,
breaking the deserializing process on the JS side.

This patch addresses the issue by converting the value stored in the
custom fields to a proper time object. Then the serializers, when
rendering JSON, will output a proper ISO8601 formatted string.