FIX: when liking a post as anon it should be liked once logged in
This used to work but was "broken" when we refactored the post stream
into a glimmer component.

This wasn't tested and is a fairly "niche" feature but since it was easy
to fix and test, it's back!

The idea is that when you like a post as anon, we store it in the
`KeyValueStore` and then present you with the login / signup
modal/form/page.

Then when you reload the topic as a logged in user, we will
automagically like the post for you.