FIX: Don’t create empty event dates in calendar
Currently, there’s an edge case in the calendar plugin: when a recurring
event (typically a daily one) ends on the next day, it can happen that we
create a `event_date` record with no `starts_at` attribute.

This is because `Event#calculate_next_date` doesn’t check the result
from `RRuleGenerator.generate`.

This patch addresses the issue simply by checking the value of
`RRuleGenerator.generate` and returns early if the value is `nil`.