FIX: Make sure theme site setting cache is set on theme import (#33697)
This fixes an issue identified in specs where the theme site setting
cache was not being set correctly when importing themes, leading
to an incomplete list of settings being sent down to the client
via the `themeSiteSettingOverrides` preload store.

For example, a theme might have something like this:

```
496=>{:enable_welcome_banner=>false}}
```

Which is missing the `search_experience` theme site setting.

We can address this properly by refreshing the SiteSetting theme
site setting cache after importing themes, ensuring that all
properties are correctly set.