FIX: ensures event builder correctly sets the description
There might be a slight chance that this spec is flakey but for some reason setting the time was making my spec to hand, I need to dig this later.

For now this commit is going to correctly ensure that when creating a new event through the form builder we set the description as the body of the event and not a property. This change allows us to have line breaks in the description and is now correctly tested.