DEV: Try fix flaky specs from theme site settings
We are seeing these flakes a lot, either the theme site setting
values don't exist or are the wrong values. Earlier, I didn't
see this TestSetup class which runs before each test, so let's
try resetting the settings and the cache before each test to make
sure things are clean.