FIX: Show localization for regionless locale if they exist (#33702)
Currently if a post is written in Japanese (`ja`) and has been
translated to English (`en`), a British (`en_GB`) user will see the
Japanese post.

WIth this PR we normalize user locales so the best match would show
(`en` first, then `en_GB`, or `pt` and `pt_BR`).

Note: `LocaleNormalizer` is from discourse-ai and will be removed there.