FIX: Don't delete keys for unsuccessfully hydrated subtypes (#33615)
`rest_serializer` option allows the backend to send a single response
containing the main object plus all its related data, and the frontend
automatically reconstructs the full object graph with proper
relationships. It is using recurrence, and will automatically detect all
id/ids and try to hydrate them into the corresponding array using the
first part as the model type name.

The problem is that the previous algorithm would delete the source array
regardless of whether the hydration was successful. This caused arrays
such as sidebar_category_ids to be automatically deleted when
currentUser was part of the model.

This commit fixes the above problem by detecting whether the hydration
was successful