FIX: Show bulk removal of expired invites  (#33716)
An existing codepath to bulk delete invites exists, but does not show
up. This commit fixes it and also fixes the old code's bug.

Specifically,
- `destroy_all_expired` endpoint only acts on the `current_user` which
is absolutely counter-intuitive, when a person visits another person’s
profile and is able to see the button, clicking it would delete their
own invites and not those of the profile they are visiting.
- In `UserInvitedShowController` the removedAll state is never set to
the original state, which results in the link being displayed in the
list when the new link expires, but the batch deletion cannot be
displayed unless the page is refreshed. It is replaced by a toast.