DEV: Capture unhandled JS errors in Playwright logs
Before this change, even if we were failing rspec tests due to
JS errors, they were not being shown in the "JS LOGS" backed
by the playwright logger in the failure message. Sometimes the
unhandled JS errors are the _cause_ of the failure, so it
is important to see them.