DEV: Reduce surface-area of ThemeJavascriptCompiler API
Extracted from #33103

The goal here is for `append_tree` to be the only interface for introducing new source files, and for all compilation to be deferred until the final `compile!` call. This change will allow us to swap out the compiler internals without any changes to call sites.

Also updates the theme tests to create a single dedicated bundle, instead of re-using and overriding parts of the normal theme bundle.