FIX: ensures we show an error when description is too long (#33750)
Sadly we can't do much better ATM, and the post will still be created.
The lifecycles of the post and the event are separated. A post can be
valid while the event is invalid, resulting in a post being created but
not the event. At least we will now show the correct description length
error instead of a 500.