FIX: default channel name to topic title (#33754)
We added few weeks ago the possibility to create a channel with the post
event, however if the event name was not set we could create an event
with no name. This change ensures we will use the topic title in this
case, as we do for the event.

Channel name max length is using the topic title max length so that
shouldn't be a problem.

---------

Co-authored-by: Jarek Radosz <jradosz@gmail.com>