DEV: Update migration to add index to run first (#33782)
This updates the timestamp of the `AddIndexToAiTopicsEmbeddings`
migration to `20241230153301` so that it runs before all recent
migrations. This is required because the lack of this index is causing
core migrations to fail as queries against the `ai_topics_embeddings`
table are running for too long and causing lock timeouts to happen.

This is a follow-up to cc77e73cfdc381a2cdce18e9e1afa869c16dcd16