DEV: Fix theme site setting flakiness part 2
In certain scenarios in the specs, we are ending up in
a situation where the default theme has been changed,
but `Theme.user_theme_ids` does not contain that theme's
ID, even though it should have the default theme ID. This
means that the default theme isn't selected as the
application_controller `@theme_id`, which is usually used
to get the correct theme site setting values from the cache.

This is happening because we need to clear the Theme cache
every time this changes. Usually this would happen via
`theme.set_default!`, but this isn't consistently used,
and we don't want to end up in a bad state when people use
`SiteSetting.default_theme_id=` directly.

This commit fixes the flakiness by moving `Theme.expire_site_cache!`
to the main site setting change tracker, this way it's always called
whenever the setting changes.