DEV: allow floatkit components to accept autoPlacement as a placement strategy (#33783)
`flip` and `autoPlacement` are [mutually exclusive
middleware](https://floating-ui.com/docs/autoplacement#conflict-with-flip)
dealing with placement. While our current Floatkit design uses `flip` by
default, there may be use cases where we want to specify using
`autoPlacement` instead.

Terming this as a choice of `visibilityOptimizer` which borrows from
floating-ui terminology of middleware that deals with `visibility
optimization`, while deconflicting with an existing `placementStrategy`
option in select-kit that is unrelated to this.