FIX: ensures autogrid works with French spacing
French spacing refers to the extra space that comes before many punctuation marks in French.

It's also common in other languages: Italian, Spanish, Catalan, ...

Our regexes were only considering the case without space and as a result auto grid would not work in French.

This regex based parsing is brittle as proven here, but this commit only focuses on fixing the bug and ensuring this case is tested for future improvements.