FIX: correctly update last message bus id for threads and channels
Given we cache the channels and threads we also store the last message bus id returned by the serializer initially. We need to ensure we update this id when we receive a message otherwise if you leave a channel and come back to it we will load the state from the id stored in the cache initially.

threads were doing it but we were not using the correct parameter.

This commit will fix multiple bugs, as for example notices that would replay each time you visit channel. We also took this opportunity to clear a notice once you leave a channel even if you didn't click dismiss.