FEATURE: create new topic while viewing restricted category or tag (#33495)
When visiting a restricted category or tag where the current user does
not have permission to create a topic, we should still allow the user to
write a new topic under a different category or tag. This is especially
important when using the **new topic** button from the sidebar.

When clicking the **new topic** button when the user does not have
permission, the composer category defaults to:

1) Subcategory (if
`SiteSetting.default_subcategory_on_read_only_category` is `true` and
subcategory exists)
2) Default site category (if `SiteSetting.default_composer_category` has
a category id value)
3) Default category chooser value of `Category...` prompting the user to
select a category from the dropdown