UX: on plugin index, unlink preinstall, simplify install banner (#33910)
The link we were using only applied to newly moved plugins, so wasn't
the best fit for all of them. This unlinks it.

I've also removed the plugin install banner and inlined it in the page
description. The blue banner was kind of intense for a message that's
always visible on the page. We need to retain the class there so it can
be hidden for sites on our hosting.

Before:
<img width="1802" height="988" alt="image"
src="https://github.com/user-attachments/assets/d101e0e0-9476-4077-9815-3427e0db69f4"
/>


After:
<img width="1778" height="804" alt="image"
src="https://github.com/user-attachments/assets/e04309e5-1d14-4f97-b43f-7f6ba7b8a483"
/>