FIX: Consistently apply chat HTML + body classes
We had a bug where the chat drawer CSS classes were
still being applied to the body/html elements even
when in full screen chat mode. This happens because
we are hadding the classes in the drawer component,
but weren't removing them when transitioning to
the full page chat route.

Might be nice to have a way to "deregister" classes
for body/html to avoid this.