DEV: Fix flaky test when editing category logo image (#33961)
Introduced in 497bfec3a959cfdc359003506ccdf20476284ebb.

This test was flaky due to the page having multiple uploaders on the same page, so the `have_css(".uploaded-image-preview.input-xxlarge")` check will always be true for the entire page. It would sometimes work correctly depending on if the uploader managed to finish quickly enough.