FIX: Composer mode was not being remembered on new PM (#33962)
If a user had the composer mode set to RTE or Markdown,
and started a new post or topic, the composer remembered
this preference and used it in the editor.

However, when the user navigated to their messages and started
a New Message, the composer mode was not being remembered.

This was happening because we fetch the user when opening the
inbox, and this ends up replacing the currentCurrent user service
object in memory. We only added `composition_mode` to the
`CurrentUserSerializer` originally, so we lost this property when
fetching the user again.

The fix here is to just add `composition_mode` to the
`UserOptionSerializer`