DEV: Fix a flaky deprecated settings test (#33996)
Some of the deprecated settings tests set our site settings to be global settings. When the settings are reloaded between tests (see `stub_deprecated_settings!()`), the setting loader (`SiteSettingExtension.setting()`) will check to see if there's a global setting that would override the site setting, and use that value from previous tests.

To fix this, we need to un-global the site setting after the test finishes.