DEV: Configure `skip_default_job_logging` for Sidekiq
By default, Sidekiq logs the `start` and `done` messages at the `info` log level
when a sidekiq job executes. These log messages are not necessary in
production and is bloating our log files. Even in other environments, I
don't see how the messages are useful so we are going to just disable
this.