DEV: Clear serializers cache between tests (#34008)
In our serializers, we have a method `cache_fragment` that we use for
caching chunks of JSON that are expensive to generate. We currently
don't clear this cache between tests, which causes tests to become flaky
since a cached fragment from one test can leak to another test and make
it fail.