FIX: Initialize mobileMode earlier
Early initializers (e.g. chat-setup) were checking mobile mode before the lib was initialized. This moves the init into the root of the module, so it's definitely ready before anything accesses it.

Also updates the chat-emoji-button `displayed` property to be a getter, so that it updates dynamically in 'viewport based mobile mode'.