FIX: Event notification translation for predefined attendance
When users were sent a notification when they were signed up to
an event with predefined attendance, the translation key was misssing,
since we originally needed to append `_html` to the key, but don't
anymore.

Handled this broken translation and used the eventName instead
of description in the notification, since the latter is not always
set and was not passed to I18n anyway.