FIX: Show category localization selector for non-supported locales when localization exists (#34025)
There's a bug now where if a category localization exists for an
unsupported locale, its locale would show up defaulting to the first
item in the dropdown. By right, the dropdown only shows locales set in
`SiteSetting.content_localization_supported_locales`.

This commit makes sure that localization's locale is added to the dropdown.