FEATURE: theme default option in user interface
Before the "theme default" option appeared conditionally, which caused confusion. For example, it was shown when the theme was using a color scheme that was not user-selectable. If the scheme was selectable, then a specific scheme like "Merigold" was preselected. If the theme changed default schemes, that change was not reflected on the user interface.

Therefore, it would be better to always have the "Theme default" option, which would have `-1` id. It means that the user's color scheme will always follow theme defaults.