DEV: Debug flaky ai_bot homepage system test
This commit reverts af2f94cca9bd76fde14055d3087103f90a7c5c1b since it is
hiding the root cause of the problem by manually cleaning up database
records in a system test. This should not be required as the database
transaction should be rolled back at the end of every test.