DEV: adds matchTriggerWidth/matchTriggerMinWidth d-menu
The `@matchTriggerWidth` and `@matchTriggerMinWidth` options have been added on `<DMenu />` to give more control on the size of the content. They will allow to respectively set the width and minWidth of the content. This is useful if you want your trigger to have the same width than your content for example.

It defaults to false

Usage:

```gjs
<DMenu @label="a long label" @matchTriggerWidth={{true}} style="width:200px;">content</DMenu>

<DMenu @label="a long label" @matchTriggerMinWidth={{true}} style="width:200px;">content</DMenu>
```