PERF: limit the number of upcoming dates generated
If you have a lot of recurring events, and particularly when recurring every day we would generate data for each of these days for a year, which means ~365 entries, multiplied by your number of events, you would quickly ended with a very large json payload.

This commit ensures we never return more than 31 upcoming dates (31 to ensure we can show a full month of events in the month view).

This was a known issue and this fix is not a definitive solution as it comes with a massive limitation: when clicking next we won't be showing you future events. This is mostly an issue for daily events, as monthly event will still allow you to see about 3 years of events which should be enough for most cases.

The ideal solution is to load only the events (and dates) we show on the screen instead of trying to load the future state from the start. For various reasons this might be a complicated change and Im writing this commit to fix pathological cases which could prevent the page to load.